+++ scripts/module.js
@@ -1,7 +1,7 @@
 /* globals
-Hooks,
 game,
-Dialog
+Hooks
+Token
 */
 "use strict";
 
@@ -12,8 +12,13 @@ import { registerGeometry } from "./geometry/registration.js";
 
 import { targetTokenHook, combatTurnHook, dnd5ePreRollAttackHook, midiqolPreambleCompleteHook } from "./cover.js";
 import { registerLibWrapperMethods, patchHelperMethods } from "./patching.js";
-import { registerSettings, getSetting, setSetting, SETTINGS, updateConfigStatusEffects, settingsCache } from "./settings.js";
-import { registerElevationAdditions } from "./elevation.js";
+import {
+  registerSettings,
+  getSetting,
+  SETTINGS,
+  updateConfigStatusEffects,
+  updateSettingHook,
+  renderSettingsConfigHook } from "./settings.js";
 
 // Rendering configs
 import { renderDrawingConfigHook } from "./renderDrawingConfig.js";
@@ -38,6 +43,8 @@ import { Area2d } from "./Area2d.js";
 import { CoverCalculator } from "./CoverCalculator.js";
 import { ConstrainedTokenBorder } from "./ConstrainedTokenBorder.js";
 
+import { Area3dPopout, area3dPopoutData } from "./Area3dPopout.js";
+
 import * as los from "./visibility_los.js";
 
 // Ignores Cover
@@ -50,10 +57,8 @@ import {
 // Other self-executing hooks
 import "./changelog.js";
 
-Hooks.once("init", async function() {
+Hooks.once("init", function() {
   registerGeometry();
-
-  registerElevationAdditions();
   registerLibWrapperMethods();
   patchHelperMethods();
   addDND5eCoverFeatFlags();
@@ -83,6 +88,9 @@ Hooks.once("init", async function() {
       IgnoresCoverSimbuls
     },
 
+    Area3dPopout,
+    area3dPopoutData,
+
     debug: DEBUG
   };
 
@@ -96,8 +104,8 @@ Hooks.once("setup", async function() {
   // Replace topZ method for tokens
   // Do this here so that it can override method from other modules, like EV.
   Object.defineProperty(Token.prototype, "topE", {
-      get: tokenTopElevation,
-      configurable: true
+    get: tokenTopElevation,
+    configurable: true
   });
 });
 
@@ -109,13 +117,15 @@ Hooks.once("setup", async function() {
  * Returns half the height if the token is prone.
  */
 function tokenTopElevation() {
-  if ( !MODULES_ACTIVE.WALL_HEIGHT ) return this.bottomE;
+  const e = this.bottomE;
+  if ( !MODULES_ACTIVE.WALL_HEIGHT ) return e;
 
   const proneStatusId = getSetting(SETTINGS.COVER.LIVE_TOKENS.ATTRIBUTE);
   const isProne = (proneStatusId !== "" && this.actor)
     ? this.actor.effects.some(e => e.getFlag("core", "statusId") === proneStatusId) : false;
 
-  return isProne ? this.losHeight * 0.5 : this.losHeight;
+  const height = this.losHeight - e;
+  return isProne ? e + (height * 0.5) : this.losHeight;
 }
 
 /**
@@ -194,65 +204,8 @@ function updateTokenHook(document, change, options, userId) { // eslint-disable-
  */
 Hooks.on("renderDrawingConfig", renderDrawingConfigHook);
 
-
-/**
- * Wipe the settings cache on update
- */
-Hooks.on("updateSetting", updateSettingHook);
-
-function updateSettingHook(document, change, options, userId) {  // eslint-disable-line no-unused-vars
-  const [module, ...arr] = document.key.split(".");
-  const key = arr.join("."); // If the key has periods, multiple will be returned by split.
-  if ( module === MODULE_ID && settingsCache.has(key) ) settingsCache.delete(key);
-}
-
-/**
- * A hook event that fires whenever an Application is rendered. Substitute the
- * Application name in the hook event to target a specific Application type, for example "renderMyApplication".
- * Each Application class in the inheritance chain will also fire this hook, i.e. "renderApplication" will also fire.
- * The hook provides the pending application HTML which will be added to the DOM.
- * Hooked functions may modify that HTML or attach interactive listeners to it.
- *
- * @event renderApplication
- * @category Application
- * @param {Application} application     The Application instance being rendered
- * @param {jQuery} html                 The inner HTML of the document that will be displayed and may be modified
- * @param {object} data                 The object of data used when rendering the application
- */
+// Note: Settings hooks
+// Settings manipulations to hide unneeded settings
+// Wipe the settings cache on update
 Hooks.on("renderSettingsConfig", renderSettingsConfigHook);
-
-/**
- * Register listeners when the settings config is opened.
- */
-function renderSettingsConfigHook(application, html, data) {
-  util.log("SettingsConfig", application, html, data);
-
-  const tvSettings = html.find(`section[data-tab="${MODULE_ID}"]`);
-  if ( !tvSettings || !tvSettings.length ) return;
-
-  const losAlgorithm = getSetting(SETTINGS.LOS.ALGORITHM);
-  const coverAlgorithm = getSetting(SETTINGS.COVER.ALGORITHM);
-
-  const displayArea = losAlgorithm === SETTINGS.LOS.TYPES.POINTS ? "none" : "block";
-  const inputLOSArea = tvSettings.find(`input[name="${MODULE_ID}.${SETTINGS.LOS.PERCENT_AREA}"]`);
-  const divLOSArea = inputLOSArea.parent().parent();
-  divLOSArea[0].style.display = displayArea;
-
-  const [displayCoverTriggers, displayCenterCoverTrigger] = coverAlgorithm === SETTINGS.COVER.TYPES.CENTER_CENTER
-    ? ["none", "block"] : ["block", "none"];
-
-  const inputCenter = tvSettings.find(`select[name="${MODULE_ID}.${SETTINGS.COVER.TRIGGER_CENTER}"]`);
-  const inputLow = tvSettings.find(`input[name="${MODULE_ID}.${SETTINGS.COVER.TRIGGER_PERCENT.LOW}"]`);
-  const inputMedium = tvSettings.find(`input[name="${MODULE_ID}.${SETTINGS.COVER.TRIGGER_PERCENT.MEDIUM}"]`);
-  const inputHigh = tvSettings.find(`input[name="${MODULE_ID}.${SETTINGS.COVER.TRIGGER_PERCENT.HIGH}"]`);
-
-  const divInputCenter = inputCenter.parent().parent();
-  const divInputLow = inputLow.parent().parent();
-  const divInputMedium = inputMedium.parent().parent();
-  const divInputHigh = inputHigh.parent().parent();
-
-  if ( divInputCenter.length ) divInputCenter[0].style.display = displayCenterCoverTrigger;
-  if ( divInputLow.length ) divInputLow[0].style.display = displayCoverTriggers;
-  if ( divInputMedium.length ) divInputMedium[0].style.display = displayCoverTriggers;
-  if ( divInputHigh.length ) divInputHigh[0].style.display = displayCoverTriggers;
-}
+Hooks.on("updateSetting", updateSettingHook);