+++ scripts/visibility_los.js
@@ -1,8 +1,9 @@
 /* globals
-Token,
 canvas,
-ClockwiseSweepPolygon,
-CONFIG
+CONFIG,
+LimitedAnglePolygon,
+PointSourcePolygon,
+Token
 */
 "use strict";
 
@@ -104,7 +105,10 @@ export function _createPolygonVisionSource(config) {
   if ( this._losCache[config.type] ) return this._losCache[config.type];
 
   const origin = { x: this.data.x, y: this.data.y };
-  const poly = CONFIG.Canvas.losBackend.create(origin, config);
+
+  // See PointSource.prototype._createPolygon
+  const polygonClass = CONFIG.Canvas.polygonBackends[config.type];
+  const poly = polygonClass.create(origin, config);
   this._losCache[config.type] = poly;
   return poly;
 }
@@ -157,6 +161,18 @@ function drawDebugPoint(visionSource, pt, hasLOS) {
   });
 }
 
+function isConstrained(los) {
+  const boundaryShapes = los.config.boundaryShapes;
+  if ( boundaryShapes.length === 0 ) return false;
+  if ( boundaryShapes.length >= 2 ) return true;
+
+  const boundaryShape = boundaryShapes[0];
+  if ( !(boundaryShape instanceof LimitedAnglePolygon) ) return true;
+
+  return boundaryShape.radius < canvas.dimensions.maxR;
+}
+
+
 /**
  * Test a point for line-of-sight. Confirm:
  * 1. Point is on the same level as the visionSource.
@@ -176,20 +192,6 @@ function testLOSPoint(visionSource, target, test) {
 
   // If not within LOS, then we are done.
   if ( MODULES_ACTIVE.PERFECT_VISION ) {
-    function isConstrained(los) {
-      const boundaryShapes = los.config.boundaryShapes;
-      if ( boundaryShapes.length === 0 ) {
-          return false;
-      }
-      if ( boundaryShapes.length >= 2 ) {
-          return true;
-      }
-      const boundaryShape = boundaryShapes[0];
-      if ( !(boundaryShape instanceof LimitedAnglePolygon) ) {
-          return true;
-      }
-      return boundaryShape.radius < canvas.dimensions.maxR;
-    }
     if ( !isConstrained(visionSource.los) ) {
       if ( !visionSource.los.contains(pt.x, pt.y) ) return false;
     } else {
@@ -197,12 +199,10 @@ function testLOSPoint(visionSource, target, test) {
       if ( angle !== 360 ) {
         const dx = pt.x - visionSource.x;
         const dy = pt.y - visionSource.y;
-        if ( dx * dx + dy * dy > externalRadius * externalRadius ) {
-          const aMin = rotation + 90 - angle / 2;
+        if ( (dx * dx) + (dy * dy) > (externalRadius * externalRadius) ) {
+          const aMin = rotation + 90 - (angle / 2);
           const a = Math.toDegrees(Math.atan2(dy, dx));
-          if ( ((a - aMin) % 360 + 360) % 360 > angle ) {
-              return false;
-          }
+          if ( ((((a - aMin) % 360) + 360) % 360) > angle ) return false;
         }
       }
       const origin = { x: visionSource.x, y: visionSource.y };
@@ -211,9 +211,7 @@ function testLOSPoint(visionSource, target, test) {
         return false;
       }
     }
-  } else {
-    if ( !visionSource.los.contains(pt.x, pt.y) ) return false;
-  }
+  } else if ( !visionSource.los.contains(pt.x, pt.y) ) return false;
 
   // If not within the constrained token shape, then don't test.
   // Assume that unconstrained token shapes contain all test points.
@@ -226,7 +224,7 @@ function testLOSPoint(visionSource, target, test) {
 
   // Test all non-infinite walls for collisions
   if ( MODULES_ACTIVE.LEVELS ) return !CONFIG.Levels.API.testCollision(origin, pt);
-  else return !ClockwiseSweepPolygon.testCollision3d(origin, pt, { type: "sight", mode: "any", wallTypes: "limited" });
+  else return !PointSourcePolygon.testCollision3d(origin, pt, { type: "sight", mode: "any", wallTypes: "limited" });
 }
 
 /**
@@ -248,7 +246,7 @@ function testLOSArea(visionSource, target, test) {
   const config = {
     type: "sight",
     liveTokensBlock: false,
-    deadTokensBlock: false,
+    deadTokensBlock: false
   };
 
   if ( DEBUG.forceLiveTokensBlock ) config.liveTokensBlock = true;
@@ -277,7 +275,7 @@ function testLOSArea3d(visionSource, target, test) {
   const config = {
     type: "sight",
     liveTokensBlock: false,
-    deadTokensBlock: false,
+    deadTokensBlock: false
   };
 
   if ( DEBUG.forceLiveTokensBlock ) config.liveTokensBlock = true;